requirements.txt
streamlit>=1.10.0
flask>=2.0.0
requests>=2.25.0
apscheduler>=3.7.0
gitpython>=3.1.0
pytest>=6.2.0
flake8>=4.0.0
mypy>=0.910